Create handler to retry rejected messages
This change introduces a retry handler. If a message is rejected for any
reason, the same message will be processed again for a given number of
times. The default number of retries is 5.

Do not foget to include the routing key for your dlx queue:
`x-dead-letter-routing-key` as demonstrated in the comments.