Fix handler interface and use real exponential backoff
We were using the wrong arity for the `reject` method in the
`ExponentialBackoffHandler`. When a message is being requeued, `Sneakers`
calls it with 4 parameters instead of 3. I changed the tests for this
handler to actually test all the possible responses (`reject`, `error`,
`timeout` and `requeue`) to make sure it always works.

I also removed the `delays` options and added real exponential
backoff, so the only thing the worker can specify is the maximum number
of retries, with the `max_retries` options, that uses a default of 25.