Ability to configure backoff function
Add new option in the worker setup, called `backoff_function`. You can
pass any `call`able object to it (a method, a lambda, a class that
implements `call`, etc.) and it will be called receiving the current
attempt number.

The default backoff function, for example, is a lambda like that:

```ruby
-> (attempt_number) { (attempt_number + 1) ** 2 }
```