Support fanout exchanges for `RetryHandler` and `DeadLetterHandler`
We need to create the dlx as a `topic` exchange. We are currently just using the
same type as the primary exchange, so here's what would happen with a `fanout`:

1) Say we have 2 queues, `q1` and `q2`, that are bound to the `f` exchange, that
is defined as a `fanout`
2) When the handler creates the dlx and error queues, we would have:
`q1.error` bound to `f.dlx`
`q2.error` bound to `f.dlx`

3) When a message on `q1` fails, it it's published to the dead-letter exchange
(`f.dlx`) using the routing key defined in the `x-dead-letter-routing-key`
argument.

4) As `fanout` exchanges will ignore this routing key, the message would end up
being sent to both `q1.error` and `q2.error`, instead of just to `q1.error`.

With this change, as `f.dlx` is always created as a `topic` exchange, the
message will be sent just to `q1.error`.