Record "rejection_reason" for the RetryHandler
This is something that we already do for the `ExponentialBackoffHandler`, so I'm
just replicating the same behaviour for the `RetryHandler`.

When a message fails because of an unhandled exception it's useful to see this
exception in the queue headers, so it's a lot easier to debug what caused the
failure.