Add tests for the rejection reason in the `ExponentialBackoffHandler`
When a message fails and is sent to the `.error` queue, we add a
`rejection_reason` property in the message's header, so it's a lot easier to
identify what caused the issue. This usually includes the exception/message that
triggered the failure.

This PR is just testing this behaviour is working as we expect.