Support for `rabbitmq-delayed-message-exchange` plugin
These changes are necessary to make this handler work when using the
`rabbitmq-delayed-message-exchange` plugin. They are:

* Create exchange using the same `type` provided in the exchange configuration:
We were always trying to create a `topic` exchange, but when an exchange is
using this plugin it will have the `x-delayed-message` type. This is will also
make the handler work with other standard types, like `direct`.

* Remove `x-delay` from message headers before publishing to the retry queues.
This is the header that `rabbitmq-delayed-message-exchange` uses to decide for
how long the message will be delayed. That's fine when we are publishing to the
original queue, but when we are handling a failure and need to publish this same
message to a retry queue, we don't want it to be delayed. This causes an even
bigger problem when the `x-delay` value is higher then the `x-expires` value for
the queue. If, for instance, we publish a message with `x-delay=2000`, and the
first retry queue has `x-expires=1000`, by the time RabbitMQ delivers this
message to the queue, it doesn't exist anymore, and we lose the message.