[Bugfix] Record reason for latest failure
This is a scenario where we'd report the wrong failure reason:

1) Message is published to queue;
2) Message fails with error `first_error`;
3) We record this failure in the `rejection_reason` header and publish this
message to the retry queues, until it reaches the error queue;
4) We can then inspect the message headers (e.g. in the management UI) and see
the `rejection_reason` to investigate what happened;
5) After the issue is fixed we shovel the message back to the working queue;
6) This time the message fails with error `second_error`;
7) As this message was just being [rejected](https://github.com/alphasights/sneakers_handlers/blob/2004126174f2d3a4d9afb05dcab0a98fb571c1f0/lib/sneakers_handlers/exponential_backoff_handler.rb#L83),
we were not able to update the message headers with the new `rejection_reason`,
so if you checked this message you would still see `first_error`.

Now instead of rejecting the original message we are acknowledging it and
directly publishing a copy of this message to the error queue with the correct
headers.