Log errors in the `retry_message` method
An exception may be raised when running this method (e.g. a
`Bunny::PreconditionFailed` when there's already a  queue defined with different
arguments), and we don't have a good way to monitor that.

Some errors may be unrecoverable and need manual intervention, so at least we
can log these errors and use a log monitoring tool to keep an eye on them.