Fix thread safety issue when creating retry queues
The exponential backoff handler creates new queues during the message
processing (since the retry queues are ephemeral) and not only during
its initialization. This is an issue because Sneakers creates one
channel per worker and shares it between threads. It is ok to consume
messages from a shared channel, but it's not thread safe to publish or
declare queues (synchronous operations) as multiple threads might try to
create retry queues at the same time.

I'm not sure how to test it, though. Concurrency hazards are random by
nature. The minimal reproduction code can be found in issue #38.